% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_blended_weight.R
\name{calculate_blended_weight_kin}
\alias{calculate_blended_weight_kin}
\title{Calculate Blended Weight for Kin Estimates}
\usage{
calculate_blended_weight_kin(main_estimate)
}
\arguments{
\item{main_estimate}{data.frame with mortality estimates, containing columns for 'type' (kin or neighbor), 'death_rate', and 'bootstrap_iter'.}
}
\value{
A single numeric value representing the blended weight for kin estimates.
}
\description{
This function calculates the blended weight to be applied to kin estimates in a mortality study.
The weight is derived based on the variance of estimates from kin and neighbor data, as well as their covariance.#'
#'
Example usage:
kin_blended_weight <- calculate_blended_weight_kin(mortality_estimates)
}
