% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cdr_kin.R
\name{calculate_cdr_kin}
\alias{calculate_cdr_kin}
\title{Estimate death rate using kin network}
\usage{
calculate_cdr_kin(death_df, survey_df, weight_col = "weights", subpop = NULL)
}
\arguments{
\item{death_df}{data.frame with deaths}

\item{survey_df}{data.frame with survey responses (and unique id to match to death_df)}

\item{subpop}{subpopulation (e.g., zone de sante)}
}
\description{
Estimate death rate using kin network
}
