% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cdr_kin_monthly.R
\name{calculate_cdr_kin_monthly}
\alias{calculate_cdr_kin_monthly}
\title{Estimate Monthly Mortality Rate for Kin}
\usage{
calculate_cdr_kin_monthly(
  death_df,
  survey_df,
  weight_col = "weights",
  subpop = NULL
)
}
\arguments{
\item{death_df}{data.frame with death records, including a 'death_month' column.}

\item{survey_df}{data.frame with survey responses, including 'num_total_kin' and a unique identifier to match with death_df.}

\item{weight_col}{string indicating the column name in survey_df used for weighting (default is "weights").}

\item{subpop}{optional vector of column names in survey_df representing subpopulations (e.g., 'zone de sante').}
}
\value{
A data frame with monthly mortality rates, both weighted and unweighted, for each subpopulation and overall.

Example usage:
estimate_mortality_kin_monthly(death_df = death_data, survey_df = survey_data, weight_col = "weight_column", subpop = c("region", "age_group"))
}
\description{
Calculates monthly mortality rates for kin based on death records and survey data, optionally within subpopulations.
The function allows for weighted and unweighted estimations.
}
