% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cdr_neighbor_monthly.R
\name{calculate_cdr_neighbor_monthly}
\alias{calculate_cdr_neighbor_monthly}
\title{Estimate Monthly Mortality Rate Among Neighbors}
\usage{
calculate_cdr_neighbor_monthly(
  death_df,
  survey_df,
  weight_col = "weights",
  subpop = NULL
)
}
\arguments{
\item{death_df}{data.frame with death records, including a 'death_month' column and a 'death_relationship/neighbour' indicator.}

\item{survey_df}{data.frame with survey responses, including 'num_total_hh_neighbour' and a unique identifier to match with death_df.}

\item{weight_col}{string indicating the column name in survey_df used for weighting.}

\item{subpop}{optional vector of column names in survey_df representing subpopulations.}
}
\value{
A data frame with monthly mortality rates among neighbors, both weighted and unweighted, for each subpopulation and overall.

examples
estimate_mortality_neighbor_monthly(death_df = death_data, survey_df = survey_data, weight_col = "weight_column", subpop = c("region", "age_group"))
}
\description{
This function calculates monthly mortality rates among neighbors based on death records and survey data. It can be tailored to specific subpopulations. The function supports both weighted and unweighted calculations.
}
