% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststrat_weights.R
\name{calculate_poststrat_weights}
\alias{calculate_poststrat_weights}
\title{Calculate Poststratification Weights for Survey Data}
\usage{
calculate_poststrat_weights(weighting_targets, survey_df)
}
\arguments{
\item{weighting_targets}{data.frame representing the target population distribution across demographic categories (like age, gender, and geographic zone).}

\item{survey_df}{data.frame with survey data, including demographic variables matching those in weighting_targets.}
}
\value{
A data frame identical to survey_df but with an additional column of poststratification weights, indicating how much each respondent's answers should count towards the overall survey results.

example
poststratified_survey <- calculate_poststrat_weights(weighting_targets = population_data, survey_df = survey_data)
}
\description{
This function calculates poststratification weights for survey data based on demographic targets. It adjusts the survey sample to be more representative of the target population by considering key demographic variables such as age, gender, and geographic zones.
}
\details{
Poststratification is a common technique in survey analysis to correct for sampling biases and make the sample more representative of the target population. This function aligns the survey sample distribution with known population distributions in different demographic categories.
}
