% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cdr_comprehensive.R
\name{compute_cdr_comprehensive}
\alias{compute_cdr_comprehensive}
\title{Estimate death rates}
\usage{
compute_cdr_comprehensive(
  death_df,
  survey_df,
  weight_col = "weight_poststrat",
  bootstrap = NA,
  monthly = FALSE,
  subpopulation = NULL,
  HH = F,
  blended_weight_kin = 0.5014031,
  weight_targets = weighting_targets,
  weight_type = "poststrat"
)
}
\arguments{
\item{death_df}{data.frame with death records.}

\item{survey_df}{data.frame with survey data.}

\item{weight_col}{character, column name of statistical person-level weights}

\item{bootstrap}{numeric, number of bootstrap samples to generate; NA for no bootstrapping.}

\item{monthly}{boolean, whether to calculate estimates on a monthly basis.}

\item{subpopulation}{optional, a vector indicating subpopulation groups for detailed analysis.}

\item{blended_weight_kin}{numeric, the weight to be applied to kin estimates to estimate the blended result.}
}
\value{
A data frame containing exposure and death estimates for kin, neighbor, and blended data types, including bootstrap iterations if applicable.

Example usage:
results <- estimate_death_rate(death_df = death_records, survey_df = survey_data, bootstrap = 100, monthly = TRUE, subpopulation = c("urban", "rural"), blended_weight_kin = 0.5014031)
}
\description{
This function computes exposure and death estimates for kin and neighbors, incorporating
options for bootstrapping, monthly calculations, and subpopulation analysis. It also calculates a
blended estimate combining kin and neighbor data, weighted by a given factor.
#'
}
