% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mortality_kin.R
\name{estimate_mortality_kin}
\alias{estimate_mortality_kin}
\title{Estimate mortality kin}
\usage{
estimate_mortality_kin(
  death_df,
  survey_df,
  weight_col = "weights",
  subpop = NULL
)
}
\arguments{
\item{death_df}{data.frame with deaths}

\item{survey_df}{data.frame with survey responses (and unique id to match to death_df)}

\item{subpop}{subpopulation (e.g., zone de sante)}
}
\value{
The negative log likelihood of parameter estimates given observed data
}
\description{
Estimate mortality rate for kin
}
