% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ipw_weights.R
\name{generate_ipw_weights}
\alias{generate_ipw_weights}
\title{Calculate Poststratification Weights for Survey Data}
\usage{
generate_ipw_weights(weighting_targets_micro, survey_df)
}
\arguments{
\item{survey_df}{data.frame with survey data, including demographic variables matching those in weighting_targets.}

\item{weighting_targets}{data.frame representing the target population distribution across demographic categories (like age, gender, and geographic zone).}
}
\value{
A data frame identical to survey_df but with an additional column of poststratification weights, indicating how much each respondent's answers should count towards the overall survey results.

example
generate_raking_weights(weighting_targets = population_data, survey_df = survey_data)
}
\description{
This function calculates  weights for survey data based on demographic targets.
}
